requirements.txt
Django==5.0.8
pillow==10.4.0
djangorestframework==3.15.2
psycopg2==2.9.9
djangorestframework-simplejwt==5.3.1
djoser==2.2.3